pull out closure_address_table code
