+++ src/EXTLLVM.cpp
@@ -68,16 +68,18 @@
 #include <random>
 #include <fstream>
 #include "stdarg.h"
-#include "EXTLLVM.h"
-#include "EXTThread.h"
-#include "UNIV.h"
-#include "SchemeFFI.h"
-#include "TaskScheduler.h"
-#include "Scheme.h"
-#include "pcre.h"
-#include "OSC.h"
-#include "math.h"
-#include "BranchPrediction.h"
+
+#include <EXTLLVM.h>
+#include <EXTClosureAddressTable.h>
+#include <EXTThread.h>
+#include <UNIV.h>
+#include <SchemeFFI.h>
+#include <TaskScheduler.h>
+#include <Scheme.h>
+#include <pcre.h>
+#include <OSC.h>
+#include <math.h>
+#include <BranchPrediction.h>
 
 #ifdef _WIN32
 #include <malloc.h>
@@ -430,88 +432,7 @@ EXPORT float imp_rand2_f(float Start, float Limit)
 
 ///////////////////////////////////
 
-///////////////////////////////////////////////////////////////////////
-// This here for Extempore Compiler Runtime.
-// This is temporary and needs to replaced with something sensible!
-struct closure_address_table
-{
-    uint64_t id;
-    char* name;
-    uint32_t offset;
-    char* type;
-    struct closure_address_table* next;
-};
 
-EXPORT closure_address_table* get_address_table(const char* name, closure_address_table* table)
-{
-  while(table)
-    {
-      if(strcmp(table->name,name)) return table;
-      table = table->next;
-    }
-  // printf("Unable to locate %s in closure environment a\n",name);
-  return 0;
-}
-
-EXPORT uint32_t get_address_offset(uint64_t id, closure_address_table* table)
-{
-    while(table)
-    {
-      // printf("%p name: %s\ntablename: %s\n\n", name, name, table->name);
-      if(table->id == id) {
-        // printf("in %s returning offset %d from %s\n",table->name,table->offset,name);
-        return table->offset;
-      }
-      table = table->next;
-    }
-    // printf("Unable to locate %" PRIu64 " in closure environment b\n",id);
-    return 0;
-}
-
-EXPORT bool check_address_exists(uint64_t id, closure_address_table* table)
-{
-    do {
-        if (table->id == id) {
-            return true;
-        }
-        table = table->next;
-    } while (table);
-    return false;
-}
-
-EXPORT bool check_address_type(uint64_t id, closure_address_table* table, const char* type)
-{
-  while(table)
-    {
-      if(table->id == id) {
-        if((strcmp(table->type,type)!=0) && (strcmp("{i8*, i8*, void (i8*, i8*)*}**",type)!=0)) {
-          printf("Runtime Type Error: bad type %s for %s. Should be %s\n",type,table->name,table->type);
-          return 0;
-        }else{
-          return 1;
-        }
-      }
-      table = table->next;
-    }
-  // printf("Unable to locate id in closure environment type: %s d\n",type);
-  return 0;
-}
-
-EXPORT closure_address_table* add_address_table(llvm_zone_t* zone, char* name, uint32_t offset, char* type, int alloctype, struct closure_address_table* table)
-{
-    struct closure_address_table* t = NULL;
-    if (alloctype == 1) {
-        t = reinterpret_cast<closure_address_table*>(malloc(sizeof(struct closure_address_table)));
-    } else {
-        t = (struct closure_address_table*) extemp::EXTLLVM::llvm_zone_malloc(zone,sizeof(struct closure_address_table));
-    }
-    t->id = string_hash(name);
-    t->name = name;
-    t->offset = offset;
-    t->type = type;
-    t->next = table;
-    return t;
-}
 
 bool llvm_check_valid_dot_symbol(scheme* sc, char* symbol) {
   char c[1024];
@@ -531,7 +452,7 @@ bool llvm_check_valid_dot_symbol(scheme* sc, char* symbol) {
   return y != sc->NIL;
 }
 
-static char* get_address_type(uint64_t id, closure_address_table* table)
+static char* get_address_type(uint64_t id, extemp::ClosureAddressTable::closure_address_table* table)
 {
     while (table)
     {
@@ -582,7 +503,7 @@ pointer llvm_scheme_env_set(scheme* _sc, char* sym)
   size_t*** closur = (size_t***) p();
   size_t** closure = *closur;
   //uint32_t** closure = (uint32_t**) cptr_value(pair_car(args));
-  closure_address_table* addy_table = (closure_address_table*) *(closure+0);
+  extemp::ClosureAddressTable::closure_address_table* addy_table = (extemp::ClosureAddressTable::closure_address_table*) *(closure+0);
   // check address exists
   if(!check_address_exists(id, addy_table)) {
     ascii_error();
@@ -598,7 +519,7 @@ pointer llvm_scheme_env_set(scheme* _sc, char* sym)
   }
   char* eptr = (char*) *(closure+1);
   char* type = get_address_type(id,addy_table);
-  uint32_t offset = get_address_offset(id,addy_table);
+  uint32_t offset = extemp::ClosureAddressTable::get_address_offset(id,addy_table);
 
   //printf("type: %s  offset: %d\n",type, offset);
 
@@ -820,7 +741,7 @@ void initLLVM()
     }
 
       // tell LLVM about some built-in functions
-            EE->updateGlobalMapping("get_address_offset", (uint64_t)&get_address_offset);
+            EE->updateGlobalMapping("get_address_offset", (uint64_t)&extemp::ClosureAddressTable::get_address_offset);
             EE->updateGlobalMapping("string_hash", (uint64_t)&string_hash);
             EE->updateGlobalMapping("swap64i", (uint64_t)&swap64i);
             EE->updateGlobalMapping("swap64f", (uint64_t)&swap64f);
